requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
slack-bolt==1.18.0
